First development commit on version 1.1 of the UniCC Parser Generator, now working with the current libphorward 0.18 development version using plist-enabled regular expression structures.
This version is no stable or well tested, but it can compile itself.
The regex anomaly feature in p_main.c has been disabled because it currently
runs into an infinite loop.

I think there is much to do on this "flagship" to prove its future.
The libphorward approach of an library-embedded parser generator will still be
continued but UniCC must stay alive and made future-ready (and maybe slim down a little bit... the character-based parsing feature is too oversized).